refactor(cli): extract lifecycle command family into cli/commands/lifecycle.ts
